Create CODE-1 for code for ultrasonic based vacuum cleaner project with Arduino board with L293D Motor Driver
This code uses the Ultrasonic library to communicate with an ultrasonic sensor and the Arduino analogWrite and digitalWrite functions to control the movement of the vacuum cleaner using an L293D motor driver. It initializes the ultrasonic sensor and the motor driver in the setup function, then continuously measures the distance to the nearest object in the loop function. If the distance is less than the cleaning width of the vacuum cleaner plus the obstacle threshold, the vacuum cleaner stops. If the distance is greater than this value, the vacuum cleaner continues moving at the set speed.

Note that this is just sample code and will likely need to be modified to fit the specific needs of your project. You may need to import additional libraries or define additional functions to control the movement of the vacuum cleaner.